Move search tickets function to utils
