+++ server/api/utils.py
@@ -8,9 +8,14 @@ import asyncio
 import logging
 from asyncio import ensure_future
 from functools import wraps
+from korail2 import NeedToLoginError
+from starlette.concurrency import run_in_threadpool
+from time import strftime
 from traceback import format_exception
 from typing import Any, Callable, Coroutine, Optional, Union
-from starlette.concurrency import run_in_threadpool
+
+from . import crud
+from .database import SessionLocal
 
 
 NoArgsNoReturnFuncT = Callable[[], None]
@@ -104,3 +109,43 @@ def repeat_every(
         return wrapped
 
     return decorator
+
+
+def search_tickets(korail, dep, arr):
+    """Search for available tickets"""
+    trains = korail.search_train_allday(
+        dep=dep,
+        arr=arr,
+        date=strftime("%Y%m%d"),
+        time="000000",
+    )
+    # FIXME: matching algorithm, O(NM). could be improved to O(N+M).
+    with SessionLocal() as db_session:
+        tickets = crud.get_tickets(
+            db_session=db_session,
+        )
+        for train in trains:
+            for ticket in tickets:
+                if (
+                    train.dep_name == ticket.departure_station and
+                    train.arr_name == ticket.arrival_station and
+                    train.dep_date == ticket.date and
+                    train.dep_time > ticket.departure_base and
+                    train.arr_time < ticket.arrival_limit and
+                    ticket.reserved == False
+                ):
+                    try:
+                        crud.mark_ticket_reserved(db_session, ticket.id)
+                        korail.login(
+                            korail_id=ticket.korail_id,
+                            korail_pw=ticket.korail_pw,
+                        )
+                        korail.reserve(train)
+                    except NeedToLoginError:
+                        crud.mark_ticket_reserved(
+                            db_session,
+                            ticket.id,
+                            False
+                        )
+                    finally:
+                        korail.logout()
